Latest attempt at D3 project (wednesday)
Having trouble getting radii to show up as functions of the data. Currently there's a sqaure root equation in there since a linear scale of circles isn't visually accurate ... but even when I was doing a straight-up "r" as a function of the data, it wouldn't show up. I made sure to convert the demographic data into numbers.